+++ ricaun.Revit.DB/Revit/Commands/Command.cs
@@ -1,6 +1,7 @@
 ﻿using Autodesk.Revit.Attributes;
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
+using System;
 using System.Diagnostics;
 using System.Linq;
 
@@ -19,7 +20,7 @@ namespace ricaun.Revit.DB.Revit.Commands
                 // Element exists in the document
                 var familyCollector = document.Select<Family>();
 
-                Debug.Assert(familyCollector.FirstElement().Id == familyCollector.FirstElementId());
+                Assert(familyCollector.FirstElement().Id == familyCollector.FirstElementId());
             }
 
             {
@@ -27,25 +28,25 @@ namespace ricaun.Revit.DB.Revit.Commands
                 var familySymbolCollector = document.Select<FamilySymbol>();
                 var familyInstanceCollector = document.Select<FamilyInstance>();
 
-                Debug.Assert(familySymbolCollector.FirstElement() is null);
-                Debug.Assert(familySymbolCollector.FirstElementId() == ElementId.InvalidElementId);
+                Assert(familySymbolCollector.FirstElement() is null);
+                Assert(familySymbolCollector.FirstElementId() == ElementId.InvalidElementId);
 
-                Debug.Assert(familyInstanceCollector.FirstElement() is null);
-                Debug.Assert(familyInstanceCollector.FirstElementId() == ElementId.InvalidElementId);
+                Assert(familyInstanceCollector.FirstElement() is null);
+                Assert(familyInstanceCollector.FirstElementId() == ElementId.InvalidElementId);
             }
 
             {
                 var elementCollector = document.Select<Element>();
                 var selectElementCollector = document.SelectElements<Element>();
 
-                Debug.Assert(elementCollector.FirstElementId() == selectElementCollector.FirstElementId());
-                Debug.Assert(elementCollector.GetElementCount() == selectElementCollector.GetElementCount());
+                Assert(elementCollector.FirstElementId() == selectElementCollector.FirstElementId());
+                Assert(elementCollector.GetElementCount() == selectElementCollector.GetElementCount());
 
                 var elementTypeCollector = document.Select<ElementType>();
                 var selectElementTypeCollector = document.SelectElementTypes<ElementType>();
 
-                Debug.Assert(elementTypeCollector.FirstElementId() == selectElementTypeCollector.FirstElementId());
-                Debug.Assert(elementTypeCollector.GetElementCount() == selectElementTypeCollector.GetElementCount());
+                Assert(elementTypeCollector.FirstElementId() == selectElementTypeCollector.FirstElementId());
+                Assert(elementTypeCollector.GetElementCount() == selectElementTypeCollector.GetElementCount());
             }
 
             {
@@ -54,20 +55,20 @@ namespace ricaun.Revit.DB.Revit.Commands
                 var elementIds = document.GetElementIds<Element>();
                 var firstElementId = document.GetFirstElementId<Element>();
 
-                Debug.Assert(elements.Count() == elementIds.Count());
-                Debug.Assert(elements.First().Id == firstElementId);
-                Debug.Assert(elements.First().Id == elementIds.First());
-                Debug.Assert(firstElement.Id == firstElementId);
+                Assert(elements.Count() == elementIds.Count());
+                Assert(elements.First().Id == firstElementId);
+                Assert(elements.First().Id == elementIds.First());
+                Assert(firstElement.Id == firstElementId);
 
                 var elementTypes = document.GetElementTypes<ElementType>();
                 var firstElementType = document.GetFirstElementType<ElementType>();
                 var elementTypeIds = document.GetElementTypeIds<ElementType>();
                 var firstElementTypeId = document.GetFirstElementTypeId<ElementType>();
 
-                Debug.Assert(elementTypes.Count() == elementTypeIds.Count());
-                Debug.Assert(elementTypes.First().Id == firstElementTypeId);
-                Debug.Assert(elementTypes.First().Id == elementTypeIds.First());
-                Debug.Assert(firstElementType.Id == firstElementTypeId);
+                Assert(elementTypes.Count() == elementTypeIds.Count());
+                Assert(elementTypes.First().Id == firstElementTypeId);
+                Assert(elementTypes.First().Id == elementTypeIds.First());
+                Assert(firstElementType.Id == firstElementTypeId);
             }
 
             {
@@ -76,32 +77,69 @@ namespace ricaun.Revit.DB.Revit.Commands
                 var id = element.Id;
                 var filterId = BuiltInParameter.ID_PARAM.Filter(id);
                 var elementFilterId = document.GetFirstElementType(filterId);
-                System.Console.WriteLine(id);
+                Console.WriteLine(id);
 
-                Debug.Assert(element.Id == elementFilterId.Id);
-
-                //var typeId = element.GetTypeId();
-                //var filterTypeId = BuiltInParameter.SYMBOL_ID_PARAM.Filter(typeId);
-                //var elementFilterTypeId = document.GetFirstElementType(filterTypeId);
-                //System.Console.WriteLine(typeId);
-
-                //Debug.Assert(element.Id == elementFilterTypeId.Id);
+                Assert(element.Id == elementFilterId.Id);
 
                 var name = element.Name;
-                var filterTypeName = BuiltInParameter.ALL_MODEL_TYPE_NAME.Filter<FilterStringEquals>(name);
+                var filterTypeName = BuiltInParameter.ALL_MODEL_TYPE_NAME.Filter(name);
                 var elementFilterName = document.GetFirstElementType(filterTypeName);
-                System.Console.WriteLine(name);
+                Console.WriteLine(name);
 
-                Debug.Assert(element.Id == elementFilterName.Id);
+                Assert(element.Id == elementFilterName.Id);
 
                 var buildInCategory = element.Category.Id.GetBuiltInCategory();
                 var elementCategory = document.GetFirstElementType(buildInCategory);
-                System.Console.WriteLine(buildInCategory);
+                Console.WriteLine(buildInCategory);
 
-                Debug.Assert(element.Id == elementCategory.Id);
+                Assert(element.Id == elementCategory.Id);
             }
 
+            {
+                var filterInverted = BuiltInParameter.SYMBOL_ID_PARAM.Rule(ElementId.InvalidElementId)
+                    .ToElementFilter(true);
+
+                var element = document.GetFirstElement(filterInverted);
+
+                var typeId = element.GetTypeId();
+                var filterTypeId = BuiltInParameter.SYMBOL_ID_PARAM.Filter(typeId);
+                var elementFiltered = document.GetFirstElement(filterTypeId);
+                System.Console.WriteLine(typeId);
+
+                Debug.Assert(element.Id == elementFiltered.Id);
+            }
+
+            {
+                var element = document.GetFirstElement<ProjectInfo>();
+                Console.WriteLine(element.Id);
+
+                Assert(element.Id == document.ProjectInformation.Id);
+
+                var elementProjectInformation = document.GetFirstElement(BuiltInCategory.OST_ProjectInformation);
+                Console.WriteLine(elementProjectInformation.Id);
+
+                Assert(elementProjectInformation.Id == document.ProjectInformation.Id);
+
+
+                var filterProjectName = BuiltInParameter.ALL_MODEL_INSTANCE_COMMENTS.Filter<FilterStringContains>(string.Empty);
+                var elementNoProjectName = document.GetFirstElement(filterProjectName);
+                Console.WriteLine(elementNoProjectName.Id);
+
+                Assert(elementNoProjectName.Id == ElementId.InvalidElementId);
+
+
+            }
+
+
             return Result.Succeeded;
+        }
+
+        public static void Assert(bool condition, [System.Runtime.CompilerServices.CallerLineNumber] int lineNumber = 0)
+        {
+            if (!condition)
+            {
+                throw new System.Exception($"Assertion failed at line {lineNumber}");
+            }
         }
     }
 }